functionality for removing conversations

+++ FinalProject/FinalProject/Controllers/ViewControllers/SettingsViewController.swift
@@ -58,6 +58,12 @@ class SettingsViewController: UIViewController {
     }
     
     @IBAction func deleteButtonTapped(_ sender: Any) {
+        guard let uid = UserController.shared.currentUser?.uuid else { return }
+        MessageController.shared.deleteUser(with: uid) { (success) in
+            if success {
+                print("Message user deleted successfully.")
+            }
+        }
         deleteUserAlert()
     }
